refactor(writeback): extract writeback enqueuer/outbox to internal/writeback/ (PKG-4d)
Move writeback_enqueuer.go and writeback_outbox.go from internal/server/ to
internal/writeback/. Replace *Server receiver/field accesses with explicit
constructor parameters. Update all call sites in internal/server/.

Also move writeback_outbox_test.go to internal/writeback/outbox_test.go
so it can access the package-internal outboxPrefix constant.

Refs: PKG-4d
